+++ packages/Webkul/UI/src/DataGrid/DataGrid.php
@@ -8,10 +8,11 @@ use Illuminate\Support\Str;
 use Webkul\UI\DataGrid\Traits\ProvideBouncer;
 use Webkul\UI\DataGrid\Traits\ProvideCollection;
 use Webkul\UI\DataGrid\Traits\ProvideExceptionHandler;
+use Webkul\UI\DataGrid\Traits\ProvideExport;
 
 abstract class DataGrid
 {
-    use ProvideBouncer, ProvideCollection, ProvideExceptionHandler;
+    use ProvideBouncer, ProvideCollection, ProvideExceptionHandler, ProvideExport;
 
     /**
      * Set index columns, ex: id. Default index is `id`.
@@ -104,9 +105,9 @@ abstract class DataGrid
     /**
      * Final result of the datagrid program that is collection object.
      *
-     * @var array
+     * @var object
      */
-    protected $collection = [];
+    protected $collection;
 
     /**
      * Parsed value of the url parameters.
@@ -454,31 +455,5 @@ abstract class DataGrid
         $this->formatCollection();
 
         return response()->json($this->prepareData());
-    }
-
-    /**
-     * Export data.
-     *
-     * @return object
-     */
-    public function export()
-    {
-        if ($this->export) {
-            $this->init();
-
-            $this->addColumns();
-
-            $this->prepareTabFilters();
-
-            $this->prepareActions();
-
-            $this->prepareMassActions();
-
-            $this->prepareQueryBuilder();
-
-            return $this->getCollection();
-        }
-
-        return [];
     }
 }
