+++ packages/Webkul/UI/src/Resources/views/export/temp.blade.php
@@ -1,16 +1,12 @@
 <table>
     <thead>
         <tr>
-            @foreach ($columns as $key => $value)
-                @php
-                    $title =  $value == 'increment_id' ? 'order_id' : $value;
-                @endphp
-
-                <th>{{ $title }}</th>
+            @foreach ($columns as $column)
+                <th>{{ $column }}</th>
             @endforeach
         </tr>
     </thead>
-    
+
     <tbody>
         @foreach ($records as $record)
             <tr>
